requirements.txt
Django>=4.0.0
Pillow>=9.0.0
easyocr>=1.7.0 